backend_core/obi_work_core/requirements.txt
pandas>=2.0.0
requests>=2.31.0
cryptography>=41.0.0
python-dotenv>=1.0.0
solana>=0.30.0
solders>=0.18.0
